Added header svg file and basic single page links structure (to be modified)
This is a commit for checking the convention of building this application during the interview task. I made it to make sure that convention how implemented my very first features in aplication is compliant with your requirements and standars. I did it on very first stage of the project so as I don't want to do the work for nothing (in case it isn't correct) and possibly start from scratch